fix: convert pazz and job scraping to the new setup
